requirements.txt
Flask
python-dotenv
requests
gunicorn
boto3